[IMP] Moved method not_authenticated_redirection_url into namespace blueprints
