+++ source/app/blueprints/rest/dashboard_routes.py
@@ -42,9 +42,9 @@ from app.models.models import TaskStatus
 from app.schema.marshables import CaseTaskSchema
 from app.schema.marshables import CaseDetailsSchema
 from app.schema.marshables import GlobalTasksSchema
-from app.blueprints.access_controls import ac_requires_case_identifier, is_authentication_oidc
+from app.blueprints.access_controls import ac_requires_case_identifier, is_authentication_oidc, \
+    not_authenticated_redirection_url
 from app.blueprints.access_controls import ac_api_requires
-from app.util import not_authenticated_redirection_url
 from app.blueprints.responses import response_error
 from app.blueprints.responses import response_success
